requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
ffpyplayer==4.5.1
filelock==3.16.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
fuzzywuzzy==0.18.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.68.0
grpcio-status==1.68.0
gTTS==2.5.3
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipython==8.12.3
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyboard==0.13.5
kiwisolver==1.4.7
langdetect==1.0.9
Levenshtein==0.26.1
llvmlite==0.43.0
loc==0.0.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.18
mistune==3.0.2
ml_dtypes==0.5.0
more-itertools==10.5.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai-whisper==20240930
opencage==3.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
playsound==1.3.0
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.8.3
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline==2.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-vlc==3.0.21203
pyttsx3==2.97
pytweening==1.2.0
pywhatkit==5.4
pywhatsapp==0.0.7
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
ratelim==0.1.6
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
screeninfo==0.8.1
selenium==4.25.0
sentencepiece==0.2.0
setuptools==75.1.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
spotipy==2.24.0
stack-data==0.6.3
sympy==1.13.2
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.20.3
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.2
transliterate==1.10.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wavio==0.0.9
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wikipedia==1.4.0
Wikipedia-API==0.7.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.11.1
yowsup2==2.5.0
